requirements.txt
# Runtime dependencies
numpy
Pillow
opencv-python
scikit-image
